add mesage routes

+++ src/app/modules/blog/blog.service.ts
@@ -1,3 +1,5 @@
+import status from 'http-status';
+import AppError from '../../error/appError';
 import { TBlog } from './blog.interface';
 import { Blog } from './blog.model';
 
@@ -13,10 +15,18 @@ export const getAllBlogsFromDB = async () => {
 };
 
 const getSingleBlogFromDB = async (id: string) => {
+  const blogId = await Blog.findById(id);
+  if (!blogId) {
+    throw new AppError(status.NOT_FOUND, 'blog id not found');
+  }
   const result = await Blog.findById(id);
   return result;
 };
 const updateBlogFromDB = async (id: string, payload: Partial<TBlog>) => {
+  const blogId = await Blog.findById(id);
+  if (!blogId) {
+    throw new AppError(status.NOT_FOUND, 'blog id not found');
+  }
   const result = await Blog.findByIdAndUpdate(id, payload, {
     new: true,
     runValidators: true,
@@ -24,6 +34,10 @@ const updateBlogFromDB = async (id: string, payload: Partial<TBlog>) => {
   return result;
 };
 const deleteBlogFromDB = async (id: string) => {
+  const blogId = await Blog.findById(id);
+  if (!blogId) {
+    throw new AppError(status.NOT_FOUND, 'blog id not found');
+  }
   const result = await Blog.findByIdAndDelete(id);
   return result;
 };
